+++ glmnetforpython/elnet.py
@@ -239,7 +239,7 @@ def elnet(
     fit["lambdau"] = alm
     fit["npasses"] = nlp_r.value
     fit["jerr"] = jerr_r.value
-    fit["dim"] = np.asarray([nvars, lmu], dtype=np.integer)
+    fit["dim"] = np.asarray([nvars, lmu], dtype=np.int64)
     fit["offset"] = is_offset
     fit["class"] = "elnet"
